current weather refactor working

+++ app/src/main/java/com/example/weatherforecast/data/models/network/Daily.kt
@@ -21,7 +21,7 @@ data class Daily(
     @SerializedName("moonset")
     val moonset: Int,
     @SerializedName("pop")
-    val pop: Double,
+    val pop: Int,
     @SerializedName("pressure")
     val pressure: Int,
     @SerializedName("rain")
